tests

+++ src/main/java/org/python/integration/object/PythonInt.java
@@ -1,9 +1,22 @@
 package org.python.integration.object;
 
-
 import java.math.BigInteger;
 
+/**
+ * Represents a Python int numbers in Java.
+ *
+ * <p>This class is a wrapper for Python int values that
+ * provides methods for safe type conversions from Java {@code long} primitives, to Java {@code int}, {@code long}, {@code BigInteger} and memory management.
+ *
+ * <p> For conversion to Java types, use
+ * {@code int} when the Python {@code int} is known to fits in 32 bits,
+ * {@code long} for 64-bit Python values, and
+ * {@code BigInteger} for values greater than 64-bit or when the size is unknown.
+ *
+ * @see IPythonObject
+ */
 public class PythonInt extends AbstractPythonObject {
+
     private PythonInt(long index, long scopeId) {
         super(index, scopeId);
     }
@@ -14,7 +27,7 @@ public class PythonInt extends AbstractPythonObject {
     }
 
     public int toJavaInt() {
-        return (int) this.toJavaLong();
+        return Math.toIntExact(this.toJavaLong());
     }
 
     public native long toJavaLong();
